Small changes proposed to enhance the code. (#8)
* [TOOLS] FSE: handle host path the same way as ibposix.c source code


* [OS] avoid warnings about signed and unsigned comparison.

This warning shows up when you turn on -Wextra compile option.


* [Strings] Make sure copied strings are 0 terminated.


* Removed unnecessary string checks

Also changed FSSTRESS_BUF_SZ to FSSTRESS_BUF_SIZE, since we use "sz" for
"String, Zero-terminated" in variable names.

* Removed unnecessary string check

* Error if pszInDir is too long

* Fix string copy checks

+++ os/linux/services/osbdev.c
@@ -682,7 +682,7 @@ static REDSTATUS FileDiskOpen(
                 fprintf(stderr, "Error: block device size (%" PRIu64 ") is smaller than requested size (%" PRIu64 ").\n", ullDevSize, (uint64_t)(ulVolSecSize * ullVolSecCount));
                 ret = -RED_EINVAL;
             }
-            else if(iSectorSize != ulVolSecSize)
+            else if(iSectorSize != (int)ulVolSecSize)
             {
                 fprintf(stderr, "Error: device sector size (%d) is different from the requested sector size (%d).\n", iSectorSize, ulVolSecSize);
                 ret = -RED_EINVAL;
@@ -754,7 +754,7 @@ static REDSTATUS FileDiskRead(
 
         result = read(pDisk->fd, pBuffer, readlen);
 
-        if(result != readlen)
+        if(result != (ssize_t)readlen)
         {
             ret = -RED_EIO;
         }
